A number of tweaks to pass CRAN checks.
Now it get by with just one NOTE---because the package is too big with all the data provided in it.  We will have to just whittle that down.  The SAMS can be gzipped down to 500 KB or so.

But, we need to use fewer loci, maybe, for fish2.  Currently it is about 7 Mb, and we'd like it to be much smaller.