explicit make mvShinyHaplot to overwrite previous shiny directory and related files;
for ui.R, the latest version of shiny the numericInput's step has internal problem where it doubles the number of input steps

server.R - make sure all indivPerDisplay and locusPerDisplay and their associate page are reset when user openning new panels and changing the number of indiv or locus per display. I also enforce that in order a violin plot to create, it has to have at least 5 points in the datatable. For per indiv and per locus panel, the datatable is pre-filtered based on the page size and page #; I did fix several warnings regarding scale_x_discrete